use single int substitution matrix remove unncessary header include

+++ src/nw_diag_avx2_256_64.c
@@ -16,7 +16,6 @@
 #include "parasail.h"
 #include "parasail/memory.h"
 #include "parasail/internal_avx.h"
-#include "parasail/matrices/blosum_map.h"
 
 #define NEG_INF (INT64_MIN/(int64_t)(2))
 
@@ -158,10 +157,10 @@ parasail_result_t* FNAME(
         __m256i vIns = vNegInf;
         __m256i vDel = vNegInf;
         __m256i vJ = vJreset;
-        const int8_t * const restrict matrow0 = &matrix->matrix[matrix->size*s1[i+0]];
-        const int8_t * const restrict matrow1 = &matrix->matrix[matrix->size*s1[i+1]];
-        const int8_t * const restrict matrow2 = &matrix->matrix[matrix->size*s1[i+2]];
-        const int8_t * const restrict matrow3 = &matrix->matrix[matrix->size*s1[i+3]];
+        const int * const restrict matrow0 = &matrix->matrix[matrix->size*s1[i+0]];
+        const int * const restrict matrow1 = &matrix->matrix[matrix->size*s1[i+1]];
+        const int * const restrict matrow2 = &matrix->matrix[matrix->size*s1[i+2]];
+        const int * const restrict matrow3 = &matrix->matrix[matrix->size*s1[i+3]];
         vNscore = _mm256_srli_si256_rpl(vNscore, 8);
         vNscore = _mm256_insert_epi64(vNscore, tbl_pr[-1], 3);
         vWscore = _mm256_srli_si256_rpl(vWscore, 8);
